requirements.txt
pandas==1.2.4
streamlit_pandas_profiling==0.1.2
streamlit==0.82.0
numpy==1.21.0
scikit_learn==0.24.2
plotly_express==0.4.1
plotly==5.1.0
wordcloud==1.8.1
nltk==3.6.2
textblob==0.15.3
findspark
pyspark
tweepy==3.10.0
python-decouple==3.4
google-cloud-translate==2.0.1